Removed old jekyll site and replaced it with bare-bones Hugo replacement.
Deleted the old jekyll site and replaced it with a Hugo site using the docsy theme.
Still needs work to format pages but sections should be in and site should "compile" and be hostable.
To start a simple server run `hugo server`, to generate html files in the `public/` directory use `hugo`.
Make sure to use hugo extended as it's required for the theme.